chore(client): tweak ui
